[Ares] Fix pre-existing lint issues: unconvert (3) and funlen (2)
- shaderarray/builder.go: remove unnecessary uint64() conversion
- cu/defaultcoalescer.go: remove 2 unnecessary int() conversions
- matrixmultiplication/mm.go: add nolint:funlen for launchKernel
- emu/cdna3/vop3a.go: add nolint:funlen for runVPKFMAF32

+++ amd/samples/runner/timingconfig/shaderarray/builder.go
@@ -517,7 +517,7 @@ func (b *Builder) buildL1VTLBs() {
 }
 
 func (b *Builder) buildL1VCaches() {
-	l1vSize := uint64(16 * mem.KB)
+	l1vSize := 16 * mem.KB
 	if b.l1vCacheSize > 0 {
 		l1vSize = b.l1vCacheSize
 	}
